removed warnings.
[[Imported from SVN: r3380]]

+++ grid/io/file/dgfparser/test/main.cc
@@ -68,7 +68,7 @@ try {
     IteratorType endit = gridPart.end<0>();
     for (IteratorType iter=gridPart.begin<0>(); iter!=endit; ++iter) {
       std::vector<double>& param = grid.parameters(*iter);
-      assert(param.size() == grid.nofParameters(0));
+      assert( (int) param.size() == grid.nofParameters(0));
       for (size_t i=0; i<param.size(); i++) {
         // std::cout << param[i] << " ";
         eldat[index.index(*iter)*param.size()+i] = param[i];
@@ -85,7 +85,7 @@ try {
     for (IteratorType iter=gridPart.begin<GridType::dimension>();
          iter!=endit; ++iter) {
       std::vector<double>& param = grid.parameters(*iter);
-      assert(param.size() == grid.nofParameters(GridType::dimension));
+      assert( (int) param.size() == grid.nofParameters(GridType::dimension));
       // std::cout << (*iter).geometry()[0] << " -\t ";
       for (size_t i=0; i<param.size(); i++) {
         // std::cout << param[i] << " ";
